Tweaks to previous commits. More tweaks to IPC fixes. Merge resolution/compile fixes.
Further work on properly integrating, updating and fixing prosthetic options.

Repathed internal organs to organ/internal.

More work on full synthetics, changelings, overheating.

Working on getting organs to qdel properly and spawn properly when robotized.

Finalized some overheating values, added remaining prosthetic icons.

Finalizing the testing version of the full prosthetic bodies branch.

Added suit cyclers to the autolathe and map.

Fixing runtime errors.

Fixing errors.

Changelog.

Replacing limb and organ strings with constants.

Prevented brainless species from becoming full cyborgs.

Fixed issues with brain/MMIs renaming themselves inappropriately.

Various fixes and oversights.

+++ code/modules/mob/living/carbon/human/species/station/station.dm
@@ -152,17 +152,17 @@
 	reagent_tag = IS_SKRELL
 
 	has_limbs = list(
-		"chest" =  list("path" = /obj/item/organ/external/chest),
-		"groin" =  list("path" = /obj/item/organ/external/groin),
-		"head" =   list("path" = /obj/item/organ/external/head/skrell),
-		"l_arm" =  list("path" = /obj/item/organ/external/arm),
-		"r_arm" =  list("path" = /obj/item/organ/external/arm/right),
-		"l_leg" =  list("path" = /obj/item/organ/external/leg),
-		"r_leg" =  list("path" = /obj/item/organ/external/leg/right),
-		"l_hand" = list("path" = /obj/item/organ/external/hand),
-		"r_hand" = list("path" = /obj/item/organ/external/hand/right),
-		"l_foot" = list("path" = /obj/item/organ/external/foot),
-		"r_foot" = list("path" = /obj/item/organ/external/foot/right)
+		BP_TORSO =  list("path" = /obj/item/organ/external/chest),
+		BP_GROIN =  list("path" = /obj/item/organ/external/groin),
+		BP_HEAD =   list("path" = /obj/item/organ/external/head/skrell),
+		BP_L_ARM =  list("path" = /obj/item/organ/external/arm),
+		BP_R_ARM =  list("path" = /obj/item/organ/external/arm/right),
+		BP_L_LEG =  list("path" = /obj/item/organ/external/leg),
+		BP_R_LEG =  list("path" = /obj/item/organ/external/leg/right),
+		BP_L_HAND = list("path" = /obj/item/organ/external/hand),
+		BP_R_HAND = list("path" = /obj/item/organ/external/hand/right),
+		BP_L_FOOT = list("path" = /obj/item/organ/external/foot),
+		BP_R_FOOT = list("path" = /obj/item/organ/external/foot/right)
 		)
 
 /datum/species/diona
@@ -190,26 +190,26 @@
 	water and other radiation."
 
 	has_organ = list(
-		"nutrient channel" =   /obj/item/organ/diona/nutrients,
-		"neural strata" =      /obj/item/organ/diona/strata,
-		"response node" =      /obj/item/organ/diona/node,
-		"gas bladder" =        /obj/item/organ/diona/bladder,
-		"polyp segment" =      /obj/item/organ/diona/polyp,
-		"anchoring ligament" = /obj/item/organ/diona/ligament
+		O_NUTRIENT = /obj/item/organ/internal/diona/nutrients,
+		O_STRATA =   /obj/item/organ/internal/diona/strata,
+		O_RESPONSE = /obj/item/organ/internal/diona/node,
+		O_GBLADDER = /obj/item/organ/internal/diona/bladder,
+		O_POLYP =    /obj/item/organ/internal/diona/polyp,
+		O_ANCHOR =   /obj/item/organ/internal/diona/ligament
 		)
 
 	has_limbs = list(
-		"chest" =  list("path" = /obj/item/organ/external/diona/chest),
-		"groin" =  list("path" = /obj/item/organ/external/diona/groin),
-		"head" =   list("path" = /obj/item/organ/external/head/no_eyes/diona),
-		"l_arm" =  list("path" = /obj/item/organ/external/diona/arm),
-		"r_arm" =  list("path" = /obj/item/organ/external/diona/arm/right),
-		"l_leg" =  list("path" = /obj/item/organ/external/diona/leg),
-		"r_leg" =  list("path" = /obj/item/organ/external/diona/leg/right),
-		"l_hand" = list("path" = /obj/item/organ/external/diona/hand),
-		"r_hand" = list("path" = /obj/item/organ/external/diona/hand/right),
-		"l_foot" = list("path" = /obj/item/organ/external/diona/foot),
-		"r_foot" = list("path" = /obj/item/organ/external/diona/foot/right)
+		BP_TORSO =  list("path" = /obj/item/organ/external/diona/chest),
+		BP_GROIN =  list("path" = /obj/item/organ/external/diona/groin),
+		BP_HEAD =   list("path" = /obj/item/organ/external/head/no_eyes/diona),
+		BP_L_ARM =  list("path" = /obj/item/organ/external/diona/arm),
+		BP_R_ARM =  list("path" = /obj/item/organ/external/diona/arm/right),
+		BP_L_LEG =  list("path" = /obj/item/organ/external/diona/leg),
+		BP_R_LEG =  list("path" = /obj/item/organ/external/diona/leg/right),
+		BP_L_HAND = list("path" = /obj/item/organ/external/diona/hand),
+		BP_R_HAND = list("path" = /obj/item/organ/external/diona/hand/right),
+		BP_L_FOOT = list("path" = /obj/item/organ/external/diona/foot),
+		BP_R_FOOT = list("path" = /obj/item/organ/external/diona/foot/right)
 		)
 
 	inherent_verbs = list(
@@ -260,9 +260,13 @@
 	if(H.mind)
 		H.mind.transfer_to(S)
 
+	if(H.isSynthetic())
+		H.visible_message("<span class='danger'>\The [H] collapses into parts, revealing a solitary diona nymph at the core.</span>")
+		return
+
 	for(var/mob/living/carbon/alien/diona/D in H.contents)
 		if(D.client)
-			D.loc = H.loc
+			D.forceMove(get_turf(H))
 		else
 			qdel(D)
